+++ src/bi_operators/CollocationIntegrator.hpp
@@ -50,17 +50,18 @@
  *
  *  Calculates the diagonal elements of S as:
  *  \f[
- *  	S_{ii} = factor_ * \sqrt{\frac{4\pi}{a_i}}
+ *  	S_{ii} = factor * \sqrt{\frac{4\pi}{a_i}}
  *  \f]
  *  while the diagonal elements of D are:
  *  \f[
- *  	D_{ii} = -factor_ * \sqrt{\frac{\pi}{a_i}} \frac{1}{R_I}
+ *  	D_{ii} = -factor * \sqrt{\frac{\pi}{a_i}} \frac{1}{R_I}
  *  \f]
  */
 
 struct CollocationIntegrator
 {
-    CollocationIntegrator() : factor_(1.07) {}
+    CollocationIntegrator() : factor(1.07) {}
+    CollocationIntegrator(double f) : factor(f) {}
     ~CollocationIntegrator() {}
 
     /**@{ Single and double layer potentials for a Vacuum Green's function by collocation */
@@ -71,7 +72,7 @@ struct CollocationIntegrator
     template <typename DerivativeTraits>
     Eigen::MatrixXd singleLayer(const Vacuum<DerivativeTraits, CollocationIntegrator> & gf, const std::vector<Element> & e) const {
         return integrator::singleLayer(e,
-                pcm::bind(integrator::SI, this->factor_, 1.0, pcm::_1),
+                pcm::bind(integrator::SI, this->factor, 1.0, pcm::_1),
                 pcm::bind(&Vacuum<DerivativeTraits, CollocationIntegrator>::kernelS, gf, pcm::_1, pcm::_2));
     }
     /*! \tparam DerivativeTraits how the derivatives of the Greens's function are calculated
@@ -81,7 +82,7 @@ struct CollocationIntegrator
     template <typename DerivativeTraits>
     Eigen::MatrixXd doubleLayer(const Vacuum<DerivativeTraits, CollocationIntegrator> & gf, const std::vector<Element> & e) const {
         return integrator::doubleLayer(e,
-                                       pcm::bind(integrator::DI, this->factor_, pcm::_1),
+                                       pcm::bind(integrator::DI, this->factor, pcm::_1),
                                        pcm::bind(&Vacuum<DerivativeTraits, CollocationIntegrator>::kernelD, gf, pcm::_1, pcm::_2, pcm::_3));
     }
     /**@}*/
@@ -94,7 +95,7 @@ struct CollocationIntegrator
     template <typename DerivativeTraits>
     Eigen::MatrixXd singleLayer(const UniformDielectric<DerivativeTraits, CollocationIntegrator> & gf, const std::vector<Element> & e) const {
         return integrator::singleLayer(e,
-                pcm::bind(integrator::SI, this->factor_, gf.epsilon(), pcm::_1),
+                pcm::bind(integrator::SI, this->factor, gf.epsilon(), pcm::_1),
                 pcm::bind(&UniformDielectric<DerivativeTraits, CollocationIntegrator>::kernelS, gf, pcm::_1, pcm::_2));
     }
     /*! \tparam DerivativeTraits how the derivatives of the Greens's function are calculated
@@ -104,7 +105,7 @@ struct CollocationIntegrator
     template <typename DerivativeTraits>
     Eigen::MatrixXd doubleLayer(const UniformDielectric<DerivativeTraits, CollocationIntegrator> & gf, const std::vector<Element> & e) const {
         return integrator::doubleLayer(e,
-                                       pcm::bind(integrator::DI, this->factor_, pcm::_1),
+                                       pcm::bind(integrator::DI, this->factor, pcm::_1),
                                        pcm::bind(&UniformDielectric<DerivativeTraits, CollocationIntegrator>::kernelD, gf, pcm::_1, pcm::_2, pcm::_3));
     }
     /**@}*/
@@ -148,7 +149,7 @@ struct CollocationIntegrator
         for (size_t i = 0; i < mat_size; ++i) {
             // Fill diagonal
             // Diagonal of S inside the cavity
-            double Sii_I = factor_ * std::sqrt(4 * M_PI / e[i].area());
+            double Sii_I = factor * std::sqrt(4 * M_PI / e[i].area());
             // "Diagonal" of Coulomb singularity separation coefficient
             double coulomb_coeff = gf.coefficientCoulomb(e[i].center(), e[i].center());
             // "Diagonal" of the image Green's function
@@ -177,9 +178,9 @@ struct CollocationIntegrator
             double area = e[i].area();
             double radius = e[i].sphere().radius;
             // Diagonal of S inside the cavity
-            double Sii_I = factor_ * std::sqrt(4 * M_PI / area);
+            double Sii_I = factor * std::sqrt(4 * M_PI / area);
             // Diagonal of D inside the cavity
-            double Dii_I = -factor_ * std::sqrt(M_PI/ area) * (1.0 / radius);
+            double Dii_I = -factor * std::sqrt(M_PI/ area) * (1.0 / radius);
             // "Diagonal" of Coulomb singularity separation coefficient
             double coulomb_coeff = gf.coefficientCoulomb(e[i].center(), e[i].center());
             // "Diagonal" of the directional derivative of the Coulomb singularity separation coefficient
@@ -205,7 +206,7 @@ struct CollocationIntegrator
     /**@}*/
 
     /// Scaling factor for the collocation formulas
-    double factor_;
+    double factor;
 };
 
 #endif // COLLOCATIONINTEGRATOR_HPP